Import minimum version, delete comments
