+++ api/src/opentrons/protocols/api_support/definitions.py
@@ -1,10 +1,10 @@
-from opentrons.protocols import types
+from .types import APIVersion
 
-MAX_SUPPORTED_VERSION = types.APIVersion(2, 8)
+MAX_SUPPORTED_VERSION = APIVersion(2, 8)
 #: The maximum supported protocol API version in this release
 
-V2_MODULE_DEF_VERSION = types.APIVersion(2, 3)
+V2_MODULE_DEF_VERSION = APIVersion(2, 3)
 #: The first API version in which we prefer the v2 module defs
 
-MINIMUM_SUPPORTED_VERSION = types.APIVersion(2, 0)
+MINIMUM_SUPPORTED_VERSION = APIVersion(2, 0)
 #: The minimum supported protocol API version in this release
