+++ api/src/opentrons/protocols/geometry/module_geometry.py
@@ -20,7 +20,7 @@ from opentrons.protocols.implementations.labware import LabwareImplementation
 
 from opentrons_shared_data import module
 from opentrons.types import Location, Point, LocationLabware
-from opentrons.protocols.types import APIVersion
+from opentrons.protocols.api_support.types import APIVersion
 from opentrons.protocols.api_support.definitions import (
     MAX_SUPPORTED_VERSION, V2_MODULE_DEF_VERSION)
 from opentrons.protocols.geometry.deck_item import DeckItem
