+++ api/tests/opentrons/protocols/advanced_control/test_transfers.py
@@ -3,7 +3,7 @@ import pytest
 import opentrons.protocol_api as papi
 from opentrons.types import Mount, TransferTipPolicy
 from opentrons.protocols.advanced_control import transfers as tx
-from opentrons.protocols.types import APIVersion
+from opentrons.protocols.api_support.types import APIVersion
 
 
 @pytest.fixture
